Part 10

+++ domain/src/main/java/com/piyush/domain/usecase/DeleteProductUseCase.kt
@@ -3,5 +3,5 @@ package com.piyush.domain.usecase
 import com.piyush.domain.repository.CartRepository
 
 class DeleteProductUseCase (private val cartRepository: CartRepository){
-    suspend fun execute(cartItemId: Int , userId: Int) = cartRepository.deleteProduct(cartItemId, userId)
+    suspend fun execute(cartItemId: Int , userId: Long) = cartRepository.deleteProduct(cartItemId, userId)
 }
